irtest: use global arrays instead of __attributes__((section)) organization
N/A

fix __attribute__((section)) syntax issues on linux and mac
